CR266 Merged 2 variant case endpoint tests. Needed new file name to avoid strange camel case issues in Git
